refactor(builder): split builder sender from server
