+++ src/builder/task.rs
@@ -1,4 +1,8 @@
-use std::{collections::HashMap, sync::Arc, time::Duration};
+use std::{
+    collections::HashMap,
+    sync::{atomic::AtomicBool, Arc},
+    time::Duration,
+};
 
 use anyhow::{bail, Context};
 use ethers::{
@@ -7,7 +11,7 @@ use ethers::{
 };
 use ethers_signers::Signer;
 use rusoto_core::Region;
-use tokio::{select, time};
+use tokio::{select, sync::mpsc, time};
 use tokio_util::sync::CancellationToken;
 use tonic::{
     async_trait,
@@ -18,10 +22,10 @@ use url::Url;
 
 use crate::{
     builder::{
-        self,
         bundle_proposer::{self, BundleProposerImpl},
+        bundle_sender::{self, BundleSender},
         sender::get_sender,
-        server::{BuilderImpl, DummyBuilder},
+        server::BuilderImpl,
         signer::{BundlerSigner, KmsSigner, LocalSigner},
         transaction_tracker::{self, TransactionTrackerImpl},
     },
@@ -152,11 +156,17 @@ impl Task for BuilderTask {
             tracker_settings,
         )
         .await?;
-        let builder_settings = builder::server::Settings {
+
+        let builder_settings = bundle_sender::Settings {
             replacement_fee_percent_increase: self.args.replacement_fee_percent_increase,
             max_fee_increases: self.args.max_fee_increases,
         };
-        let builder = Arc::new(BuilderImpl::new(
+        let manual_bundling_mode = Arc::new(AtomicBool::new(false));
+        let (send_bundle_tx, send_bundle_rx) = mpsc::channel(1);
+
+        let mut bundle_sender = BundleSender::new(
+            manual_bundling_mode.clone(),
+            send_bundle_rx,
             self.args.chain_id,
             beneficiary,
             self.args.eth_poll_interval,
@@ -166,15 +176,16 @@ impl Task for BuilderTask {
             transaction_tracker,
             provider,
             builder_settings,
-        ));
+        );
 
         let _builder_loop_guard = {
-            let builder = Arc::clone(&builder);
-            SpawnGuard::spawn_with_guard(async move { builder.send_bundles_in_loop().await })
+            SpawnGuard::spawn_with_guard(async move { bundle_sender.send_bundles_in_loop().await })
         };
 
+        let builder_server = BuilderImpl::new(manual_bundling_mode, send_bundle_tx);
+
         // gRPC server
-        let builder_server = BuilderServer::new(builder);
+        let builder_server = BuilderServer::new(builder_server);
 
         let reflection_service = tonic_reflection::server::Builder::configure()
             .register_encoded_file_descriptor_set(BUILDER_FILE_DESCRIPTOR_SET)
@@ -183,7 +194,7 @@ impl Task for BuilderTask {
         // health service
         let (mut health_reporter, health_service) = tonic_health::server::health_reporter();
         health_reporter
-            .set_serving::<BuilderServer<DummyBuilder>>()
+            .set_serving::<BuilderServer<BuilderImpl>>()
             .await;
 
         let server_handle = Server::builder()
